refactor(site): rename css-hooks.ts module
