+++ site/src/app/docs/[framework]/layout.tsx
@@ -8,7 +8,7 @@ import Logo from "@/components/Logo";
 import PageBlock from "@/components/PageBlock";
 import RadioLink from "@/components/RadioLink";
 import Typography from "@/components/Typography";
-import { css } from "@/css-hooks";
+import { css } from "@/css";
 import { exhausted } from "@/util/exhausted";
 import frameworks from "./frameworks";
 import guides from "./guides";
